fix: address code review findings for DAC expansion
Critical fixes:
- _effectiveCaps() return type uint8_t→uint16_t to preserve extended
  capability bits (HAL_CAP_MQA, LINE_DRIVER, APLL, DSD) — pre-existing
  bug now exposed by DAC drivers using bits 8-11
- Extend ESS_DAC_READY_FN table from 8→16 entries to match
  AUDIO_OUT_MAX_SINKS=16 — prevents null isReady for sink slots 8-15

Important fixes:
- Add Wire2 extern include to hal_es9069q.cpp for ESP32 compilation

2933 tests pass, zero failures.

+++ src/hal/hal_pipeline_bridge.cpp
@@ -75,8 +75,8 @@ static uint8_t _halSlotSinkCount[HAL_MAX_DEVICES];  // Number of consecutive sin
 // explicit caps (non-zero) we trust them; otherwise fall back to type-based
 // inference for backward compatibility with devices registered before the
 // capabilities field was populated.
-static uint8_t _effectiveCaps(HalDevice* dev) {
-    uint8_t caps = dev->getDescriptor().capabilities;
+static uint16_t _effectiveCaps(HalDevice* dev) {
+    uint16_t caps = dev->getDescriptor().capabilities;
     if (caps != 0) return caps;
     // Infer from device type (conservative defaults)
     switch (dev->getType()) {
